feat(storage): add file metadata during upload (#514)
