+++ .changeset/tidy-seas-swim.md
@@ -1,0 +1,5 @@
+---
+'@capacitor-firebase/storage': minor
+---
+
+feat: support `metadata` property on upload
