+++ packages/storage/ios/Plugin/Classes/Options/UploadFileOptions.swift
@@ -1,13 +1,19 @@
 import Foundation
+import Capacitor
+import FirebaseStorage
 
 @objc public class UploadFileOptions: NSObject {
     private var path: String
     private var uri: URL
+    private var metadata: StorageMetadata?
     private var callbackId: String
 
-    init(path: String, uri: URL, callbackId: String) {
+    init(path: String, uri: URL, metadata: JSObject?, callbackId: String) {
         self.path = path
         self.uri = uri
+        if let metadata = metadata {
+            self.metadata = FirebaseStorageHelper.buildStorageMetadata(metadata)
+        }
         self.callbackId = callbackId
     }
 
@@ -17,6 +23,10 @@ import Foundation
 
     func getUri() -> URL {
         return uri
+    }
+
+    func getMetadata() -> StorageMetadata? {
+        return metadata
     }
 
     func getCallbackId() -> String {
