+++ packages/storage/src/definitions.ts
@@ -292,6 +292,20 @@ export interface SettableMetadata {
 }
 
 /**
+ * @since 5.4.0
+ */
+export interface UploadMetadata extends SettableMetadata {
+  /**
+   * The base64-encoded MD5 hash of the file.
+   *
+   * Only available for Web.
+   *
+   * @since 5.4.0
+   */
+  md5Hash?: string;
+}
+
+/**
  * @since 5.3.0
  */
 export interface UploadFileOptions {
@@ -321,6 +335,12 @@ export interface UploadFileOptions {
    * @example 'file:///var/mobile/Containers/Data/Application/E397A70D-67E4-4258-236E-W1D9E12111D4/Library/Caches/092F8464-DE60-40B3-8A23-EB83160D9F9F/mountains.png'
    */
   uri?: string;
+  /**
+   * The metadata to set for the file.
+   *
+   * @since 5.4.0
+   */
+  metadata?: UploadMetadata;
 }
 
 /**
